added Get request code

+++ src/test/java/resources/Utils.java
@@ -16,15 +16,19 @@ import io.restassured.specification.RequestSpecification;
 import io.restassured.specification.ResponseSpecification;
 
 public class Utils {
-	RequestSpecification reqspec;
-	ResponseSpecification resspec;
+	public static RequestSpecification reqspec;
+	public static ResponseSpecification resspec;
 
 	public RequestSpecification requestSpec() throws IOException {
-		PrintStream ps = new PrintStream(new FileOutputStream("log.txt"));
-		reqspec = new RequestSpecBuilder().setBaseUri(getGlobalProperties("baseURL")).setContentType(ContentType.JSON)
-				.addQueryParam("key", "qaclick123").addFilter(RequestLoggingFilter.logRequestTo(ps))
-				.addFilter(ResponseLoggingFilter.logResponseTo(ps)).build();
 
+		if (reqspec == null) {// this condition is used to store mutiple logs out puts as if request specificaiton is updated everytime then the logs are also overwritten
+			PrintStream ps = new PrintStream(new FileOutputStream("log.txt"));
+			reqspec = new RequestSpecBuilder().setBaseUri(getGlobalProperties("baseURL"))
+					.setContentType(ContentType.JSON).addQueryParam("key", "qaclick123")
+					.addFilter(RequestLoggingFilter.logRequestTo(ps)).addFilter(ResponseLoggingFilter.logResponseTo(ps))
+					.build();
+			return reqspec;
+		}
 		return reqspec;
 	}
 
@@ -37,8 +41,8 @@ public class Utils {
 		Properties prop = new Properties();
 		FileInputStream fis = new FileInputStream("src/test/java/resources/Global.properties");
 		prop.load(fis);
-		
+
 		return prop.getProperty(key);
-		
+
 	}
 }
